+++ scenedetect/scene_manager.py
@@ -51,14 +51,17 @@ threshold values (or other algorithm options) are used.
 from __future__ import print_function
 from string import Template
 import math
+import os
 import logging
 
 # Third-Party Library Imports
 import cv2
 import numpy as np
+from edl import Event
 from scenedetect.platform import tqdm
 from scenedetect.platform import get_and_create_path
 from scenedetect.platform import get_aspect_ratio
+from timecode import Timecode
 
 # PySceneDetect Library Imports
 from scenedetect.frame_timecode import FrameTimecode
@@ -246,6 +249,61 @@ def write_scene_list_html(output_html_filename, scene_list, cut_list=None, css=N
     page.save(output_html_filename)
 
 
+def write_scene_list_edl(output_edl_filename, scene_list):
+    """Writes the given list of scenes to an output file handle in edl format.
+
+    Arguments:
+        output_edl_filename: filename of output edl file.
+        scene_list: List of pairs of FrameTimecodes denoting each scene's start/end FrameTimecode.
+
+    """
+
+    edl_events = _get_edl_events(output_edl_filename, scene_list)
+    header = 'TITLE: {}\nFCM: NON-DROP FRAME\n'.format(output_edl_filename)
+    with open(output_edl_filename, 'w') as edl_file:
+        edl_file.write('{}\n'.format(header))
+        for num , event in enumerate(edl_events):
+            edl_file.write('{}\n'.format(event.to_string()))
+
+
+def _get_edl_events(output_edl_filename, scene_list):
+    """Returns EDL Events.
+
+    Arguments:
+        output_edl_filename: filename of output edl file.
+        scene_list: List of pairs of FrameTimecodes denoting each scene's start/end FrameTimecode.
+
+    """
+    edl_events = []
+    edl_event_keys = ['num', 'reel', 'track', 'tr_code', 'aux', 'src_start_tc',
+                      'src_end_tc', 'rec_start_tc', 'rec_end_tc']
+    reel_name = os.path.splitext(output_edl_filename)[0]
+    fps = scene_list[0][0].get_framerate()
+    for num, scene in enumerate(scene_list, 1):
+        start_tc, end_tc = _get_timecodes(scene, fps)
+        values = [str(num).zfill(4), reel_name, 'V', '', 'C',
+                  start_tc, end_tc, start_tc, end_tc]
+        event = Event(dict(zip(edl_event_keys, values)))
+        # event.comments = ['* FROM CLIP NAME: {}'.format("")]
+        edl_events.append(event)
+    return edl_events
+
+
+def _get_timecodes(scene, fps):
+    """Returns tuple of start and end timecode.
+
+    Arguments:
+        scene: Pair of FrameTimecodes denoting each scene's start/end FrameTimecode.
+        fps: Framerate of source video.
+
+    """
+    start_tc = Timecode(fps)
+    end_tc = Timecode(fps)
+    start_tc.frames = scene[0].get_frames() + 1
+    end_tc.frames = scene[1].get_frames() + 1
+    return start_tc, end_tc
+
+
 def save_images(scene_list, video_manager, num_images=3, frame_margin=1,
                 image_extension='jpg', encoder_param=95,
                 image_name_template='$VIDEO_NAME-Scene-$SCENE_NUMBER-$IMAGE_NUMBER',
@@ -396,7 +454,7 @@ def save_images(scene_list, video_manager, num_images=3, frame_margin=1,
                     frame_im = cv2.resize(
                         frame_im, (0, 0), fx=aspect_ratio, fy=1.0,
                         interpolation=cv2.INTER_CUBIC)
-                
+
                 # Get frame dimensions prior to resizing or scaling
                 frame_height = frame_im.shape[0]
                 frame_width = frame_im.shape[1]